Added Nav Links, Fixed Nav
Fixed the nav. The problem was that the languages dropdown would take
up too much space and would put the other links on another line. Then,
when it opened, the min-width was 160, which I reduced to 80px. So,
whenever you open it, it doesn’t jump to another line either.